物品不再使用 BLoC 来进行表单验证
